+++ Informix.DB2/Tests/Linq/EnumMappingTests/Tests.Linq.EnumMappingTests.EnumMapSet2(Informix.DB2).sql
@@ -27,7 +27,7 @@ SET     @TestField = 12
 UPDATE
 	LinqDataTypes r
 SET
-	BigIntValue = @TestField
+	BigIntValue = @TestField::BigInt
 WHERE
 	r.ID = 101 AND r.BigIntValue = 11
