+++ Informix.DB2/Tests/UserTests/Issue421/Tests.UserTests.Issue421.Test1(Informix.DB2).sql
@@ -11,8 +11,8 @@ INSERT INTO BlobClass
 )
 VALUES
 (
-	@Id,
-	@BlobValue
+	@Id::Int,
+	@BlobValue::byte
 )
 
 -- Informix.DB2 Informix
@@ -32,7 +32,7 @@ SET     @BlobValue = {3,2,1}
 UPDATE
 	BlobClass t1
 SET
-	BlobValue = @BlobValue
+	BlobValue = @BlobValue::byte
 WHERE
 	t1.Id = 1
