+++ Informix.DB2/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.InsertEnum2(Informix.DB2).sql
@@ -18,8 +18,8 @@ INSERT INTO Parent
 )
 VALUES
 (
-	@ParentID,
-	@Value1
+	@ParentID::Int,
+	@Value1::Int
 )
 
 -- Informix.DB2 Informix
